fix: p2p-extractor integration test with addconnection
Since #323 we can use addconnection instead of --addnode.
This has the benefit that addconnection is an RPC call that's
executed instantly compared to --addnode that's executed once
on startup and then again every minute if the connection is not
yet established. This means, if the connection fails to be
established with --addnode (particuliary common on slower systems)
we'll need to wait for 60s before we can continue with the
integration test.

Using the RPC addconnection should make the tests more stable and
faster. Especially on slower systems / CI.